% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_grid.R
\name{OR_grid}
\alias{OR_grid}
\title{Make a color-coded grid of odds ratios}
\usage{
OR_grid(dat, x, y, OR = "OR", p.value = "p.value", alpha = 0.05,
  breaks = c(0, 0.3, 0.7, 1, 2, 5, Inf), label_size = NULL,
  coord_flip = FALSE)
}
\arguments{
\item{dat}{A data frame}

\item{x}{The value to map to the x axis}

\item{y}{The value to map to the y axis}

\item{OR}{The odds ratio}

\item{p.value}{The p-value}

\item{alpha}{The cutoff for significance}

\item{breaks}{The discrete values by which to color code ORs}

\item{label_size}{The size of the labels}

\item{coord_flip}{Whether or not to flip coordinates}
}
\description{
\code{OR_grid} Makes a grid of ORs, color-coded OR and significance
}
\seealso{
\code{\link[sggplots]{OR_grid_large}}
}
