% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_grid.R
\name{OR_grid_large}
\alias{OR_grid_large}
\title{Make a color-coded grid of odds ratios}
\usage{
OR_grid_large(dat, x = "Exposure.Variable", y = "Strata", OR = "OR",
  p.value = "p.value", alpha = 0.05, breaks = c(0, 0.3, 0.7, 1, 2, 5,
  Inf), label_size = NULL, coord_flip = FALSE)
}
\arguments{
\item{dat}{A data frame}

\item{x}{The value to map to the x axis}

\item{y}{The value to map to the y axis}

\item{OR}{The odds ratio}

\item{p.value}{The p-value}

\item{alpha}{The cutoff for significance}

\item{breaks}{The discrete values by which to color code ORs}

\item{label_size}{The size of the labels}

\item{coord_flip}{Whether or not to flip coordinates}
}
\description{
Make a color-coded grid of odds ratios
}
